# internal/http/api/api.gen.go
//go:build go1.22

// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/oapi-codegen/runtime"
)

// Defines values for AvailabilityWindowDayOfWeek.
const (
	Friday    AvailabilityWindowDayOfWeek = "friday"
	Monday    AvailabilityWindowDayOfWeek = "monday"
	Saturday  AvailabilityWindowDayOfWeek = "saturday"
	Sunday    AvailabilityWindowDayOfWeek = "sunday"
	Thursday  AvailabilityWindowDayOfWeek = "thursday"
	Tuesday   AvailabilityWindowDayOfWeek = "tuesday"
	Wednesday AvailabilityWindowDayOfWeek = "wednesday"
)

// Valid indicates whether the value is a known member of the AvailabilityWindowDayOfWeek enum.
func (e AvailabilityWindowDayOfWeek) Valid() bool {
	switch e {
	case Friday:
		return true
	case Monday:
		return true
	case Saturday:
		return true
	case Sunday:
		return true
	case Thursday:
		return true
	case Tuesday:
		return true
	case Wednesday:
		return true
	default:
		return false
	}
}

// AvailabilityResponse defines model for AvailabilityResponse.
type AvailabilityResponse struct {
	// Owner Single calendar owner profile. It is configured ahead of time and does not authenticate in v1.
	Owner   OwnerProfile         `json:"owner"`
	Windows []AvailabilityWindow `json:"windows"`
}

// AvailabilityWindow Owner-defined availability window for a specific day of week.
type AvailabilityWindow struct {
	DayOfWeek AvailabilityWindowDayOfWeek `json:"dayOfWeek"`

	// EndTime Local end time in HH:MM format.
	EndTime string `json:"endTime"`

	// StartTime Local start time in HH:MM format.
	StartTime string `json:"startTime"`
}

// AvailabilityWindowDayOfWeek defines model for AvailabilityWindow.DayOfWeek.
type AvailabilityWindowDayOfWeek string

// BadRequestError defines model for BadRequestError.
type BadRequestError struct {
	Error ErrorBody `json:"error"`
}

// Booking Confirmed booking created by a guest without registration.
type Booking struct {
	EndsAt      time.Time `json:"endsAt"`
	EventTypeId string    `json:"eventTypeId"`
	GuestEmail  string    `json:"guestEmail"`
	GuestName   string    `json:"guestName"`
	Id          string    `json:"id"`
	StartsAt    time.Time `json:"startsAt"`
}

// BookingConflictError Returned when the requested time interval overlaps with an existing booking, even if the event type is different.
type BookingConflictError struct {
	Error ErrorBody `json:"error"`
}

// BookingResponse defines model for BookingResponse.
type BookingResponse struct {
	// Booking Confirmed booking created by a guest without registration.
	Booking Booking `json:"booking"`
}

// CreateBookingRequest defines model for CreateBookingRequest.
type CreateBookingRequest struct {
	EventTypeId string    `json:"eventTypeId"`
	GuestEmail  string    `json:"guestEmail"`
	GuestName   string    `json:"guestName"`
	StartsAt    time.Time `json:"startsAt"`
}

// CreateEventTypeRequest defines model for CreateEventTypeRequest.
type CreateEventTypeRequest struct {
	Description     string `json:"description"`
	DurationMinutes int32  `json:"durationMinutes"`
	Title           string `json:"title"`
}

// CreatedBookingResponse defines model for CreatedBookingResponse.
type CreatedBookingResponse struct {
	// Booking Confirmed booking created by a guest without registration.
	Booking Booking `json:"booking"`
}

// CreatedEventTypeResponse defines model for CreatedEventTypeResponse.
type CreatedEventTypeResponse struct {
	// EventType Bookable event type configured by the calendar owner.
	EventType EventType `json:"eventType"`
}

// ErrorBody defines model for ErrorBody.
type ErrorBody struct {
	Code    string `json:"code"`
	Message string `json:"message"`
}

// EventType Bookable event type configured by the calendar owner.
type EventType struct {
	Description     string `json:"description"`
	DurationMinutes int32  `json:"durationMinutes"`
	Id              string `json:"id"`
	Title           string `json:"title"`
}

// EventTypeListResponse defines model for EventTypeListResponse.
type EventTypeListResponse struct {
	Items []EventType `json:"items"`
}

// NotFoundError defines model for NotFoundError.
type NotFoundError struct {
	Error ErrorBody `json:"error"`
}

// OwnerProfile Single calendar owner profile. It is configured ahead of time and does not authenticate in v1.
type OwnerProfile struct {
	DisplayName string `json:"displayName"`
	Id          string `json:"id"`
	Timezone    string `json:"timezone"`
}

// Slot Computed free slot for a specific event type.
type Slot struct {
	DurationMinutes int32     `json:"durationMinutes"`
	EndsAt          time.Time `json:"endsAt"`
	EventTypeId     string    `json:"eventTypeId"`
	StartsAt        time.Time `json:"startsAt"`
}

// SlotListResponse defines model for SlotListResponse.
type SlotListResponse struct {
	Items []Slot `json:"items"`
}

// UpcomingBookingItem Confirmed booking created by a guest without registration.
type UpcomingBookingItem = Booking

// UpcomingBookingsResponse defines model for UpcomingBookingsResponse.
type UpcomingBookingsResponse struct {
	Items []UpcomingBookingItem `json:"items"`
}

// UpdateAvailabilityRequest defines model for UpdateAvailabilityRequest.
type UpdateAvailabilityRequest struct {
	Windows []AvailabilityWindow `json:"windows"`
}

// UpdatedAvailabilityResponse defines model for UpdatedAvailabilityResponse.
type UpdatedAvailabilityResponse struct {
	Availability AvailabilityResponse `json:"availability"`
}

// PublicEventTypesApiListSlotsParams defines parameters for PublicEventTypesApiListSlots.
type PublicEventTypesApiListSlotsParams struct {
	From time.Time `form:"from" json:"from"`
	To   time.Time `form:"to" json:"to"`
}

// PublicBookingsApiCreateJSONRequestBody defines body for PublicBookingsApiCreate for application/json ContentType.
type PublicBookingsApiCreateJSONRequestBody = CreateBookingRequest

// OwnerAvailabilityApiUpdateAvailabilityJSONRequestBody defines body for OwnerAvailabilityApiUpdateAvailability for application/json ContentType.
type OwnerAvailabilityApiUpdateAvailabilityJSONRequestBody = UpdateAvailabilityRequest

// OwnerEventTypesApiCreateJSONRequestBody defines body for OwnerEventTypesApiCreate for application/json ContentType.
type OwnerEventTypesApiCreateJSONRequestBody = CreateEventTypeRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (POST /bookings)
	PublicBookingsApiCreate(w http.ResponseWriter, r *http.Request)

	// (GET /event-types)
	PublicEventTypesApiList(w http.ResponseWriter, r *http.Request)

	// (GET /event-types/{eventTypeId}/slots)
	PublicEventTypesApiListSlots(w http.ResponseWriter, r *http.Request, eventTypeId string, params PublicEventTypesApiListSlotsParams)

	// (GET /owner/availability)
	OwnerAvailabilityApiGetAvailability(w http.ResponseWriter, r *http.Request)

	// (PUT /owner/availability)
	OwnerAvailabilityApiUpdateAvailability(w http.ResponseWriter, r *http.Request)

	// (GET /owner/bookings/upcoming)
	OwnerBookingsApiListUpcoming(w http.ResponseWriter, r *http.Request)

	// (GET /owner/event-types)
	OwnerEventTypesApiList(w http.ResponseWriter, r *http.Request)

	// (POST /owner/event-types)
	OwnerEventTypesApiCreate(w http.ResponseWriter, r *http.Request)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// PublicBookingsApiCreate operation middleware
func (siw *ServerInterfaceWrapper) PublicBookingsApiCreate(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PublicBookingsApiCreate(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PublicEventTypesApiList operation middleware
func (siw *ServerInterfaceWrapper) PublicEventTypesApiList(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PublicEventTypesApiList(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PublicEventTypesApiListSlots operation middleware
func (siw *ServerInterfaceWrapper) PublicEventTypesApiListSlots(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "eventTypeId" -------------
	var eventTypeId string

	err = runtime.BindStyledParameterWithOptions("simple", "eventTypeId", r.PathValue("eventTypeId"), &eventTypeId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "eventTypeId", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params PublicEventTypesApiListSlotsParams

	// ------------- Required query parameter "from" -------------

	if paramValue := r.URL.Query().Get("from"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "from"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", false, true, "from", r.URL.Query(), &params.From, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "from", Err: err})
		return
	}

	// ------------- Required query parameter "to" -------------

	if paramValue := r.URL.Query().Get("to"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "to"})
		return
	}

	err = runtime.BindQueryParameterWithOptions("form", false, true, "to", r.URL.Query(), &params.To, runtime.BindQueryParameterOptions{Type: "string", Format: "date-time"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "to", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PublicEventTypesApiListSlots(w, r, eventTypeId, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// OwnerAvailabilityApiGetAvailability operation middleware
func (siw *ServerInterfaceWrapper) OwnerAvailabilityApiGetAvailability(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.OwnerAvailabilityApiGetAvailability(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// OwnerAvailabilityApiUpdateAvailability operation middleware
func (siw *ServerInterfaceWrapper) OwnerAvailabilityApiUpdateAvailability(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.OwnerAvailabilityApiUpdateAvailability(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// OwnerBookingsApiListUpcoming operation middleware
func (siw *ServerInterfaceWrapper) OwnerBookingsApiListUpcoming(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.OwnerBookingsApiListUpcoming(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// OwnerEventTypesApiList operation middleware
func (siw *ServerInterfaceWrapper) OwnerEventTypesApiList(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.OwnerEventTypesApiList(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// OwnerEventTypesApiCreate operation middleware
func (siw *ServerInterfaceWrapper) OwnerEventTypesApiCreate(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.OwnerEventTypesApiCreate(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("POST "+options.BaseURL+"/bookings", wrapper.PublicBookingsApiCreate)
	m.HandleFunc("GET "+options.BaseURL+"/event-types", wrapper.PublicEventTypesApiList)
	m.HandleFunc("GET "+options.BaseURL+"/event-types/{eventTypeId}/slots", wrapper.PublicEventTypesApiListSlots)
	m.HandleFunc("GET "+options.BaseURL+"/owner/availability", wrapper.OwnerAvailabilityApiGetAvailability)
	m.HandleFunc("PUT "+options.BaseURL+"/owner/availability", wrapper.OwnerAvailabilityApiUpdateAvailability)
	m.HandleFunc("GET "+options.BaseURL+"/owner/bookings/upcoming", wrapper.OwnerBookingsApiListUpcoming)
	m.HandleFunc("GET "+options.BaseURL+"/owner/event-types", wrapper.OwnerEventTypesApiList)
	m.HandleFunc("POST "+options.BaseURL+"/owner/event-types", wrapper.OwnerEventTypesApiCreate)

	return m
}